refactor: split files for support multi strategies
