+++ src/audioDownloader/strategies/webApiGetAllGeneratingUrlsData/consts.ts
@@ -1,0 +1,9 @@
+export const GET_AUDIO_DATA_ERROR_MESSAGE =
+  "Audio downloader. WEB API. Can not get getGeneratingAudioUrlsDataFromIframe due to timeout";
+export const INCORRECT_FETCH_MEDIA_MESSAGE =
+  "Audio downloader. WEB API. Incorrect response on fetch media url";
+export const CANT_FETCH_MEDIA_MESSAGE =
+  "Audio downloader. WEB API. Can not fetch media url";
+export const CANT_GET_ARRAY_BUFFER_MESSAGE =
+  "Audio downloader. WEB API. Can not get array buffer from media url";
+export const textDecoder = new TextDecoder("ascii");
